refactor: shell rm checks file existance

+++ app/shell.py
@@ -1,5 +1,7 @@
 import os
+import os.path
 
 
 def rm(file):
-    os.remove(file)
+    if os.path.isfile(file):
+        os.remove(file)
